+++ crates/eips/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.1.1](https://github.com/alloy-rs/alloy/releases/tag/v2.1.1) - 2026-07-06
+## [2.2.0](https://github.com/alloy-rs/alloy/releases/tag/v2.2.0) - 2026-07-17
 
 ### Bug Fixes
 
@@ -20,12 +20,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [consensus] Add block gas limit validation ([#4072](https://github.com/alloy-rs/alloy/issues/4072))
 - [rpc-beacon-type] Extend ssz decoding to `SignedBeaconBlock` ([#4029](https://github.com/alloy-rs/alloy/issues/4029))
 - Add ssz encoding and decoding to Payload Envelopes ([#4024](https://github.com/alloy-rs/alloy/issues/4024))
 - [eips] Add EIP-7594 matching cell computation ([#3974](https://github.com/alloy-rs/alloy/issues/3974))
 
 ### Miscellaneous Tasks
 
+- [eips] Address nightly clippy lint ([#4080](https://github.com/alloy-rs/alloy/issues/4080))
+- Release 2.1.1
 - Expand zepter feature checks ([#4063](https://github.com/alloy-rs/alloy/issues/4063))
 - Release 2.1.0
 - Release 2.0.5